interchange: merge protobuf modules
The separation between the protobuf and protobuf::decode module didn't
quite pan out, since deriving scalar types from proto types is properly
part of decoding too. Just merge the modules together again.
