+++ src/dataflow-types/src/types.rs
@@ -366,14 +366,12 @@ impl DataEncoding {
             DataEncoding::Protobuf(ProtobufEncoding {
                 descriptors,
                 message_name,
-            }) => {
-                protobuf::decode::DecodedDescriptors::from_bytes(descriptors, message_name.into())?
-                    .validate()?
-                    .into_iter()
-                    .fold(RelationDesc::empty(), |desc, (name, ty)| {
-                        desc.with_named_column(name.unwrap(), ty)
-                    })
-            }
+            }) => protobuf::DecodedDescriptors::from_bytes(descriptors, message_name.into())?
+                .validate()?
+                .into_iter()
+                .fold(RelationDesc::empty(), |desc, (name, ty)| {
+                    desc.with_named_column(name.unwrap(), ty)
+                }),
             DataEncoding::Regex(RegexEncoding { regex }) => regex
                 .capture_names()
                 .enumerate()
